rename account_page to page
